Remove references of postings reader from index_sections_reader
