+++ pkg/dataobj/metastore/object.go
@@ -662,7 +662,6 @@ func (m *ObjectMetastore) IndexSectionsReader(ctx context.Context, req IndexSect
 		req.SectionsRequest.Matchers,
 		req.SectionsRequest.Predicates,
 		req.BatchSize,
-		false,
 	)
 
 	return IndexSectionsReaderResponse{Reader: reader}, nil
