+++ pkg/dataobj/metastore/postings_index_sections_reader_test.go
@@ -6,11 +6,16 @@ import (
 	"testing"
 	"time"
 
+	"github.com/apache/arrow-go/v18/arrow"
 	"github.com/apache/arrow-go/v18/arrow/array"
 	"github.com/go-kit/log"
 	"github.com/grafana/dskit/user"
 	"github.com/prometheus/prometheus/model/labels"
 	"github.com/stretchr/testify/require"
+
+	"github.com/grafana/loki/v3/pkg/dataobj"
+	"github.com/grafana/loki/v3/pkg/dataobj/sections/postings"
+	"github.com/grafana/loki/v3/pkg/dataobj/sections/streams"
 )
 
 func TestPostingsIndexSectionsReader_ReadBeforeOpenReturnsError(t *testing.T) {
@@ -257,4 +262,223 @@ func TestPostingsIndexSectionsReader_EndToEnd(t *testing.T) {
 		require.NotNil(t, rec)
 		require.Equal(t, int64(1), rec.NumRows())
 	})
+}
+
+// requirePostingsSchema asserts rec carries the postings.ReadPointersForStreams output
+// schema (9 pointer columns + InternalLabelsFieldName), proving the postings reader
+// emits the same schema as the streams+pointers reader.
+func requirePostingsSchema(t *testing.T, rec arrow.RecordBatch) {
+	t.Helper()
+
+	require.Equal(t, 10, len(rec.Schema().Fields()),
+		"postings.ReadPointersForStreams output has 9 columns + InternalLabelsFieldName")
+
+	for _, name := range []string{
+		"path.path.utf8",
+		"section.int64",
+		"pointer_kind.int64",
+		"stream_id.int64",
+		"stream_id_ref.int64",
+		"min_timestamp.timestamp",
+		"max_timestamp.timestamp",
+		"row_count.int64",
+		"uncompressed_size.int64",
+	} {
+		columnByName(t, rec, name) // fails the test if the column is absent
+	}
+}
+
+func columnByName(t *testing.T, rec arrow.RecordBatch, name string) arrow.Array {
+	t.Helper()
+	for i, f := range rec.Schema().Fields() {
+		if f.Name == name {
+			return rec.Column(i)
+		}
+	}
+	require.Failf(t, "missing column", "record batch has no column %q", name)
+	return nil
+}
+
+func buildPostingsFixture(t *testing.T) *dataobj.Object {
+	t.Helper()
+
+	builder := newFixtureBuilder(t)
+
+	_, err := builder.AppendStream(tenantID, streams.Stream{
+		ID:               1,
+		Labels:           labels.New(labels.Label{Name: "app", Value: "foo"}),
+		MinTimestamp:     now.Add(-3 * time.Hour),
+		MaxTimestamp:     now.Add(-2 * time.Hour),
+		UncompressedSize: 5,
+		Rows:             1,
+	})
+	require.NoError(t, err)
+
+	builder.ObserveLabelPosting(tenantID, postings.LabelObservation{
+		ObjectPath:       "test-path",
+		SectionIndex:     0,
+		ColumnName:       "app",
+		LabelValue:       "foo",
+		StreamID:         1,
+		Timestamp:        now.Add(-3 * time.Hour),
+		UncompressedSize: 5,
+	})
+
+	builder.PrepareBloomColumn(tenantID, "test-path", 0, "traceID", 1)
+	require.NoError(t, builder.ObserveBloomPosting(tenantID, postings.BloomObservation{
+		ObjectPath:       "test-path",
+		SectionIndex:     0,
+		ColumnName:       "traceID",
+		Value:            "abcd",
+		StreamID:         1,
+		Timestamp:        now.Add(-3 * time.Hour),
+		UncompressedSize: 5,
+	}))
+
+	return flushFixture(t, builder)
+}
+
+func buildMultiPostingsSectionsFixture(t *testing.T) *dataobj.Object {
+	t.Helper()
+
+	first := postings.NewBuilder(nil, 0, 0)
+	first.SetTenant(tenantID)
+	first.ObserveLabelPosting(postings.LabelObservation{
+		ObjectPath:       "test-path-1",
+		SectionIndex:     0,
+		ColumnName:       "app",
+		LabelValue:       "foo",
+		StreamID:         1,
+		Timestamp:        now.Add(-3 * time.Hour),
+		UncompressedSize: 5,
+	})
+
+	second := postings.NewBuilder(nil, 0, 0)
+	second.SetTenant(tenantID)
+	second.ObserveLabelPosting(postings.LabelObservation{
+		ObjectPath:       "test-path-2",
+		SectionIndex:     0,
+		ColumnName:       "app",
+		LabelValue:       "foo",
+		StreamID:         1,
+		Timestamp:        now.Add(-3 * time.Hour),
+		UncompressedSize: 5,
+	})
+
+	objBuilder := dataobj.NewBuilder(nil)
+	require.NoError(t, objBuilder.Append(first))
+	require.NoError(t, objBuilder.Append(second))
+	obj, closer, err := objBuilder.Flush()
+	require.NoError(t, err)
+	t.Cleanup(func() { _ = closer.Close() })
+	return obj
+}
+
+func buildPostingsMultiLabelFixture(t *testing.T) *dataobj.Object {
+	t.Helper()
+
+	builder := newFixtureBuilder(t)
+
+	_, err := builder.AppendStream(tenantID, streams.Stream{
+		ID: 1,
+		Labels: labels.New(
+			labels.Label{Name: "app", Value: "foo"},
+			labels.Label{Name: "team", Value: "ops"},
+		),
+		MinTimestamp:     now.Add(-3 * time.Hour),
+		MaxTimestamp:     now.Add(-2 * time.Hour),
+		UncompressedSize: 5,
+		Rows:             1,
+	})
+	require.NoError(t, err)
+
+	builder.ObserveLabelPosting(tenantID, postings.LabelObservation{
+		ObjectPath:       "test-path",
+		SectionIndex:     0,
+		ColumnName:       "app",
+		LabelValue:       "foo",
+		StreamID:         1,
+		Timestamp:        now.Add(-3 * time.Hour),
+		UncompressedSize: 5,
+	})
+	builder.ObserveLabelPosting(tenantID, postings.LabelObservation{
+		ObjectPath:       "test-path",
+		SectionIndex:     0,
+		ColumnName:       "team",
+		LabelValue:       "ops",
+		StreamID:         1,
+		Timestamp:        now.Add(-3 * time.Hour),
+		UncompressedSize: 5,
+	})
+
+	builder.PrepareBloomColumn(tenantID, "test-path", 0, "traceID", 1)
+	require.NoError(t, builder.ObserveBloomPosting(tenantID, postings.BloomObservation{
+		ObjectPath:       "test-path",
+		SectionIndex:     0,
+		ColumnName:       "traceID",
+		Value:            "abcd",
+		StreamID:         1,
+		Timestamp:        now.Add(-3 * time.Hour),
+		UncompressedSize: 5,
+	}))
+
+	return flushFixture(t, builder)
+}
+
+func buildPostingsStreamIDCollisionFixture(t *testing.T) *dataobj.Object {
+	t.Helper()
+
+	builder := postings.NewBuilder(nil, 0, 0)
+	builder.SetTenant(tenantID)
+
+	builder.ObserveLabelPosting(postings.LabelObservation{
+		ObjectPath:       "obj-a",
+		SectionIndex:     0,
+		ColumnName:       "app",
+		LabelValue:       "foo",
+		StreamID:         1,
+		Timestamp:        now.Add(-3 * time.Hour),
+		UncompressedSize: 5,
+	})
+	builder.ObserveLabelPosting(postings.LabelObservation{
+		ObjectPath:       "obj-b",
+		SectionIndex:     0,
+		ColumnName:       "app",
+		LabelValue:       "bar",
+		StreamID:         1,
+		Timestamp:        now.Add(-3 * time.Hour),
+		UncompressedSize: 5,
+	})
+
+	objBuilder := dataobj.NewBuilder(nil)
+	require.NoError(t, objBuilder.Append(builder))
+	obj, closer, err := objBuilder.Flush()
+	require.NoError(t, err)
+	t.Cleanup(func() { _ = closer.Close() })
+	return obj
+}
+
+// buildPostingsOnlyFixture writes only postings observations (no AppendStream /
+// ObserveLogLine), so Flush emits a postings section and no streams or pointers.
+func buildPostingsOnlyFixture(t *testing.T) *dataobj.Object {
+	t.Helper()
+
+	builder := newFixtureBuilder(t)
+
+	builder.ObserveLabelPosting(tenantID, postings.LabelObservation{
+		ObjectPath: "test-path", SectionIndex: 0, ColumnName: "app", LabelValue: "foo",
+		StreamID: 1, Timestamp: now.Add(-3 * time.Hour), UncompressedSize: 5,
+	})
+	builder.ObserveLabelPosting(tenantID, postings.LabelObservation{
+		ObjectPath: "test-path", SectionIndex: 0, ColumnName: "app", LabelValue: "bar",
+		StreamID: 2, Timestamp: now.Add(-3 * time.Hour), UncompressedSize: 5,
+	})
+
+	builder.PrepareBloomColumn(tenantID, "test-path", 0, "traceID", 1)
+	require.NoError(t, builder.ObserveBloomPosting(tenantID, postings.BloomObservation{
+		ObjectPath: "test-path", SectionIndex: 0, ColumnName: "traceID", Value: "abcd",
+		StreamID: 1, Timestamp: now.Add(-3 * time.Hour), UncompressedSize: 5,
+	}))
+
+	return flushFixture(t, builder)
 }
